+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
  * Medium changes
    - Clean up files in recent/exit-lists/ again.
+   - Retain padding-counts lines in sanitized extra-info descriptors.
 
 
 # Changes in version 1.1.2 - 2017-01-17